Clean up some warnings
